refactor: replaced useHidePassword hook
